refactor: move config sidebar to the configs directory
