+++ src/components/Configs/Sidebar/ConfigSidebar.tsx
@@ -1,14 +1,14 @@
 import { useParams } from "react-router-dom";
-import ConfigChanges from "../ConfigChanges";
-import ConfigCosts from "../ConfigCosts";
-import ConfigInsights from "../ConfigInsights";
-import Configs from "../Sidebars/configs";
-import Incidents from "../Sidebars/incidents";
-import SlidingSideBar from "../SlidingSideBar";
+import ConfigChanges from "../../ConfigChanges";
+import ConfigCosts from "../../ConfigCosts";
+import ConfigInsights from "../../ConfigInsights";
+import Configs from "../../Sidebars/configs";
+import Incidents from "../../Sidebars/incidents";
+import SlidingSideBar from "../../SlidingSideBar";
 import { ConfigDetails } from "./ConfigDetails";
 import ConfigActionBar from "./ConfigActionBar";
 import { useCallback, useState } from "react";
-import { PlaybookRunsSidePanel } from "../Playbooks/Runs/PlaybookRunsSidePanel";
+import { PlaybookRunsSidePanel } from "../../Playbooks/Runs/PlaybookRunsSidePanel";
 
 type SidePanels =
   | "ConfigDetails"
